+++ crates/ffi/CHANGELOG.md
@@ -3,7 +3,14 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.7.1] - 2024-09-23
+
+### Changed
+- Now flm `constructor` can throw exceptions.
+- Upd `flm` dependency to `0.7`
+
+[0.7.1]: https://github.com/AdguardTeam/FilterListManager/compare/ffi-0.6.3...ffi-0.7.1
 
 ## [0.6.3] - 2024-09-23